mypy

+++ src/tdg/context_managers.py
@@ -19,17 +19,18 @@ class TempDir:
             shutil.copytree(self._orig_root, self.root, dirs_exist_ok=True)
             for parent, dirs, files in os.walk(self._orig_root):
                 for file in files:
-                    file = Path(parent) / Path(file)
-                    if "pyc" not in file.suffix:
-                        rel_file = self.root / file.relative_to(self._orig_root)
-                        self._temp_files[file.stem] = rel_file
+                    filePath = Path(parent) / Path(file)
+                    if "pyc" not in filePath.suffix:
+                        rel_file = self.root / filePath.relative_to(self._orig_root)
+                        self._temp_files[filePath.stem] = rel_file
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
-        shutil.rmtree(self.root)
-        assert not self.root.exists(), FileExistsError(
-            f"self.root was not removed: {self.root}"
-        )
+        if self.root:
+            shutil.rmtree(self.root)
+            assert not self.root.exists(), FileExistsError(
+                f"self.root was not removed: {self.root}"
+            )
 
-    def __getitem__(self, key: str) -> Path:
+    def __getitem__(self, key: str) -> Optional[Path]:
         return self._temp_files.get(key.split(".")[0])
